requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
cryptography==45.0.2
distro==1.9.0
ecdsa==0.19.1
exceptiongroup==1.3.0
Flask==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.0
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
narwhals==1.40.0
numpy==2.0.2
openai==1.82.0
packaging==25.0
pandas==2.2.3
plotly==6.1.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-http-client==3.3.7
pytz==2025.2
scikit-learn==1.6.1
scipy==1.13.1
sendgrid==6.12.3
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
zipp==3.21.0
